Move filter constructors to separate file
